--- THREAD_QUEUE_CHANGES.md
@@ -1,143 +1,0 @@
-# Thread-Specific Queue Implementation
-
-## Overview
-Implemented per-thread message queuing to enable concurrent processing of messages from different threads, significantly improving efficiency and responsiveness.
-
-## Key Changes
-
-### 1. Queue Manager (`app/queue_manager.py`)
-
-**Before:**
-- Single global priority queue for all messages
-- Sequential processing across all threads
-- One message processed at a time regardless of thread
-
-**After:**
-- Per-thread priority queues (`_thread_queues: Dict[str, PriorityQueue]`)
-- Each thread has its own independent queue
-- Messages are enqueued to thread-specific queues
-- `thread_id` is now required (auto-generated by API if not provided)
-- Concurrent processing across different threads
-
-**New Methods:**
-- `get_active_threads()` - Returns set of thread IDs with pending messages
-- `has_messages(thread_id)` - Check if specific thread has pending messages
-- Updated `dequeue(thread_id)` - Dequeue from specific thread queue
-
-### 2. Worker (`app/worker.py`)
-
-**Before:**
-- Single worker loop processing one message at a time
-- All threads blocked while one message processes
-
-**After:**
-- Coordinator pattern with per-thread workers
-- Main coordinator (`_run_coordinator`) spawns workers for active threads
-- Each thread has its own worker task (`_process_thread`)
-- Concurrent processing: Thread A processing doesn't block Thread B
-- Automatic worker lifecycle management
-
-**Architecture:**
-```
-Coordinator Loop
-├── Monitors active threads
-├── Spawns worker for each active thread
-├── Cleans up completed workers
-└── Per-Thread Workers
-    ├── Thread 1 Worker → Processes Thread 1 queue
-    ├── Thread 2 Worker → Processes Thread 2 queue
-    └── Thread N Worker → Processes Thread N queue
-```
-
-### 3. API Layer (`app/api/routes.py`)
-
-**Change:**
-- Backend now auto-generates `thread_id` if not provided by client
-- Every message gets a unique thread_id for independent processing
-- Backward compatible: clients don't need to provide thread_id
-
-```python
-# Before: thread_id could be None
-thread_id = request.thread_id
-
-# After: auto-generate if not provided
-thread_id = request.thread_id or str(uuid.uuid4())
-```
-
-### 4. Models (`app/models.py`)
-
-**Updates:**
-- `ThreadMetadata.thread_id` - Required string (not Optional)
-- `ThreadSummary.thread_id` - Required string (not Optional)
-- `QueuedMessage.thread_id` - Remains Optional for internal flexibility
-
-## Benefits
-
-### Performance Improvements
-1. **Concurrent Processing**: Multiple threads process simultaneously
-2. **No Cross-Thread Blocking**: Long-running message in Thread A doesn't delay Thread B
-3. **Better Resource Utilization**: CPU cores utilized more efficiently
-
-### Scalability
-1. **Thread Isolation**: Each thread scales independently
-2. **Dynamic Worker Management**: Workers created/destroyed as needed
-3. **Memory Efficient**: Empty queues cleaned up automatically
-
-### User Experience
-1. **Faster Response Times**: No waiting for unrelated threads
-2. **Fair Processing**: Each thread progresses independently
-3. **Predictable Behavior**: Thread-specific queue position
-
-## Example Scenario
-
-**Before (Sequential):**
-```
-Thread A: [Long Message (60s)] → [Message 2]
-Thread B: [Quick Message (5s)] ← BLOCKED waiting for Thread A
-Thread C: [Quick Message (3s)] ← BLOCKED waiting for Thread B
-```
-
-**After (Concurrent):**
-```
-Thread A: [Long Message (60s)] → Processing
-Thread B: [Quick Message (5s)] → Processing concurrently → DONE in 5s
-Thread C: [Quick Message (3s)] → Processing concurrently → DONE in 3s
-```
-
-## Testing
-
-All 47 tests pass, including:
-- ✅ Thread independence verification
-- ✅ Active threads tracking
-- ✅ Concurrent processing
-- ✅ Priority ordering within threads
-- ✅ Backward compatibility (auto thread_id generation)
-
-## Migration Notes
-
-### For API Consumers
-- **No breaking changes** - thread_id is optional in requests
-- Backend auto-generates thread_id if not provided
-- Existing code continues to work without modification
-
-### For Internal Code
-- `queue_manager.enqueue()` now requires `thread_id` parameter
-- `queue_manager.dequeue()` now requires `thread_id` parameter
-- Tests updated to provide `thread_id` explicitly
-
-## Performance Characteristics
-
-- **Queue Operations**: O(log n) within each thread queue
-- **Thread Lookup**: O(1) hash-based thread queue access
-- **Worker Overhead**: One task per active thread
-- **Memory**: Separate queue per thread (minimal overhead)
-
-## Future Enhancements
-
-Potential improvements:
-1. Configurable max concurrent threads
-2. Thread priority (high-priority threads get more workers)
-3. Thread rate limiting
-4. Queue depth monitoring per thread
-5. Thread metrics and analytics
-
