+++ src/shinzo/agent/__init__.py
@@ -1,0 +1,6 @@
+"""Agent processing package for LangGraph agent integration."""
+
+from shinzo.agent.processor import AgentProcessor
+
+__all__ = ["AgentProcessor"]
+
