+++ src/shinzo/worker/__init__.py
@@ -1,0 +1,6 @@
+"""Background worker package for message processing."""
+
+from shinzo.worker.manager import Worker
+
+__all__ = ["Worker"]
+
