+++ tests/unit/test_queue_manager.py
@@ -1,5 +1,5 @@
 import pytest
-from shinzo.queue_manager import QueueManager
+from shinzo.queue import QueueManager
 from shinzo.models import Priority, MessageState
